fix(Link): rename files to project standards
