+++ packages/react/src/components/link/index.ts
@@ -1,0 +1,1 @@
+export { Link } from './link';
